+++ Sources/Paywall/Paywall/Paywall View Controller/WebEventHandler.swift
@@ -18,7 +18,9 @@ protocol WebEventHandlerDelegate: AnyObject {
   var isPresentedViewController: Bool { get }
 
   func eventDidOccur(_ paywallPresentationResult: PaywallPresentationResult)
-  func presentSafari(_ url: URL)
+  func openDeepLink(_ url: URL)
+  func presentSafariInApp(_ url: URL)
+  func presentSafariExternal(_ url: URL)
 }
 
 final class WebEventHandler: WebEventDelegate {
@@ -51,13 +53,15 @@ final class WebEventHandler: WebEventDelegate {
       delegate?.eventDidOccur(.closed)
     case .openUrl(let url):
       openUrl(url)
+    case .openUrlInSafari(let url):
+      openUrlInSafari(url)
     case .openDeepLink(let url):
       openDeepLink(url)
     case .restore:
       restorePurchases()
-    case .purchase(product: let type):
+    case .purchase(productId: let id):
       purchaseProduct(
-        withType: type,
+        withId: id,
         from: paywallResponse
       )
     case .custom(data: let customEvent):
@@ -154,11 +158,9 @@ final class WebEventHandler: WebEventDelegate {
     )
     delegate?.webView.configuration.userContentController.addUserScript(selectionScript)
 
-    // swiftlint:disable:next line_length
     let preventSelection = "var css = '*{-webkit-touch-callout:none;-webkit-user-select:none}'; var head = document.head || document.getElementsByTagName('head')[0]; var style = document.createElement('style'); style.type = 'text/css'; style.appendChild(document.createTextNode(css)); head.appendChild(style);"
     delegate?.webView.evaluateJavaScript(preventSelection)
 
-    // swiftlint:disable:next line_length
     let preventZoom: String = "var meta = document.createElement('meta');" + "meta.name = 'viewport';" + "meta.content = 'width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no';" + "var head = document.getElementsByTagName('head')[0];" + "head.appendChild(meta);"
     delegate?.webView.evaluateJavaScript(preventZoom)
   }
@@ -170,7 +172,17 @@ final class WebEventHandler: WebEventDelegate {
     )
     hapticFeedback()
     delegate?.eventDidOccur(.openedURL(url: url))
-    delegate?.presentSafari(url)
+    delegate?.presentSafariInApp(url)
+  }
+
+  private func openUrlInSafari(_ url: URL) {
+    detectHiddenPaywallEvent(
+      "openUrlInSafari",
+      userInfo: ["url": url]
+    )
+    hapticFeedback()
+    delegate?.eventDidOccur(.openedUrlInSafari(url))
+    delegate?.presentSafariExternal(url)
   }
 
   private func openDeepLink(_ url: URL) {
@@ -179,8 +191,7 @@ final class WebEventHandler: WebEventDelegate {
       userInfo: ["url": url]
     )
     hapticFeedback()
-    delegate?.eventDidOccur(.openedDeepLink(url: url))
-    // TODO: Handle deep linking
+    delegate?.openDeepLink(url)
   }
 
   private func restorePurchases() {
@@ -190,15 +201,12 @@ final class WebEventHandler: WebEventDelegate {
   }
 
   private func purchaseProduct(
-    withType productType: ProductType,
+    withId id: String,
     from paywallResponse: PaywallResponse
   ) {
     detectHiddenPaywallEvent("purchase")
     hapticFeedback()
-    let product = paywallResponse.products.first { $0.type == productType }
-    if let product = product {
-      delegate?.eventDidOccur(.initiatePurchase(productId: product.id))
-    }
+    delegate?.eventDidOccur(.initiatePurchase(productId: id))
   }
 
   private func handleCustomEvent(_ customEvent: String) {
@@ -235,12 +243,11 @@ final class WebEventHandler: WebEventDelegate {
 
   private func hapticFeedback() {
     if Paywall.isGameControllerEnabled { return }
-    
+
     if #available(iOS 13.0, *) {
       UIImpactFeedbackGenerator().impactOccurred(intensity: 0.7)
     } else {
       UIImpactFeedbackGenerator().impactOccurred()
     }
-
   }
 }