+++ Sources/Paywall/Paywall/Paywall.swift
@@ -41,7 +41,7 @@ public final class Paywall: NSObject {
 	/// Animates paywall presentation. Defaults to `true`.
   ///
   /// Set this to `false` to globally disable paywall presentation animations.
-  @available(*, deprecated, message: "Set the Presentation Style on the Superwall dashboard to No Animation instead of using this boolean.")
+  @available(*, deprecated, message: "Either set the Presentation Style on the Superwall dashboard to No Animation or, for a trigger-specific override, set presentationStyleOverride on Paywall.trigger().")
 	public static var shouldAnimatePaywallPresentation = true
 
   /// Animates paywall dismissal. Defaults to `true`.
@@ -343,6 +343,9 @@ extension Paywall: SWPaywallViewControllerDelegate {
         )
 			case .openedURL(let url):
 				Paywall.delegate?.willOpenURL?(url: url)
+      case .openedUrlInSafari(let url):
+        // TODO: Is this the right delegate call? Should we create a new one?
+        Paywall.delegate?.willOpenURL?(url: url)
 			case .openedDeepLink(let url):
 				Paywall.delegate?.willOpenDeepLink?(url: url)
 			case .custom(let string):